Create README.md
I just started coding this year and found a high interest on pool coin mining and building a site big at large here in the States as they are all pretty much in China for the most part and thought it would be nice to have a local pool to crash and a change up for others to come check out.  I'm currently learning kali Linux and ubuntu server 16.04.07 and built my first machine in December 2021 for about 40k every screw and piece put and picked by me and it being my first desktop and second computer next to my laptop I got last year.  I'm learning a new skill and having fun doing it as well I thought I would never be able to use a computer unless it had a Gui but now i rather and prefer cli over the graphical user interface any day it's fun challenging and not as hard as it looks like or I thought it would be all though it was intimidating at first glance now it's my way of life on the computer as i know it.  Looking and only looking for very experienced and knowledgeable mentors to help mold and shape my newfound freedom and career in Linux systems BareMetal and code development and network security and infrastructure using various different implemented software protocol and algos.